fix change email button at /subscription page
