+++ src/containers/Subscribtion/Home.tsx
@@ -23,7 +23,6 @@ export function SubscribeHome({ returnUrl }: { returnUrl?: string }) {
 	const [billingInterval, setBillingInterval] = useState<'year' | 'month'>('month')
 	const isSubscribed = subscription?.status === 'active'
 	const [isClient, setIsClient] = useState(false)
-	const [showEmailForm, setShowEmailForm] = useState(false)
 	const [showReturnModal, setShowReturnModal] = useState(false)
 	const [hasShownModal, setHasShownModal] = useState(false)
 
@@ -241,7 +240,6 @@ export function SubscribeHome({ returnUrl }: { returnUrl?: string }) {
 								isVerified={user?.verified}
 								isSubscribed={isSubscribed}
 								subscription={subscription}
-								onEmailChange={() => setShowEmailForm(true)}
 								getPortalSessionUrl={getPortalSessionUrl}
 							/>
 						</div>
